libia2: split out `ia2_common.h` for definitions common to `ia2.h` and `ia2_internal.h`
This lets us move all non-public definitions to `ia2_internal.h`
instead of the public `ia2.h`, namely the memory maps types.

Moreover, `ia2.h` can go back to including `ia2_internal.h`
at the top instead of the bottom, since the definitions
it needs from `ia2.h` are now in `ia2_common.h`.
