+++ runtime/libia2/include/ia2_common.h
@@ -1,0 +1,5 @@
+#pragma once
+
+/// IA2 definitions that are common to both `ia2.h` and `ia2_internal.h`.
+
+#define IA2_MAX_COMPARTMENTS 16
